\name{argparse-package}
\alias{argparse-package}
\alias{argparse}
\docType{package}
\title{
\packageTitle{argparse}
}
\description{
\packageDescription{argparse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{argparse}
\packageIndices{argparse}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{argparse}

Maintainer: \packageMaintainer{argparse}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
