% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparse.R
\name{argument_parser}
\alias{argument_parser}
\title{Parse command line arguments}
\usage{
argument_parser(..., description = NULL)
}
\arguments{
\item{...}{(list)
Each element of the list is the output of the argument_def function}

\item{description}{(character)
A description of the script to be printed when user uses --help}
}
\value{
(closure) 
     closure param (vector (character))
         A vector of command line arguments, i.e. that returned by 
         commandArgs function, to be used by script. 

     closure return (list)
     List contains default or input values for each option and argument 
     input to the closures parent function.
}
\description{
Parse command line arguments
}
