% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_r_lang_args}
\alias{parse_r_lang_args}
\title{Retreive user specified options from Rscript command line options}
\usage{
parse_r_lang_args(args)
}
\arguments{
\item{args}{character vector:
command line arguments, i.e. returned by commandArgs,
that Rscript uses to run a script}
}
\value{
output list:
 The list contains 2 key value pairs:
 * program_name = the string that proceeds command line argument '--file='
 * args = the list of user defined arguments
}
\description{
Rscipt passes the command line options specified by the user
for the script / program and other to options the R language
executable.  This function finds the options for the desired
script / program.
}
\details{
The command line arguments to an R program / script launched by
Rscript, e.g.

Rscript <script_name.r> --arg1 value --arg2 value
for running the R language executable. This function finds the
`script_name.r` string and user defined key / val pairs.
}
