% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_option}
\alias{process_option}
\title{Parse argument character vector for arg with definition arg_def}
\usage{
process_option(arg_def, args)
}
\arguments{
\item{arg_def}{list:
output from arg_def}

\item{args}{character vector:
The set of arguments to match against arg_def}
}
\value{
NA | logical | vector :
 vector type depends are arg_def specified type
}
\description{
Parse argument character vector for arg with definition arg_def
}
