% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_def.R
\name{argument_def}
\alias{argument_def}
\title{Specify command line argument}
\usage{
argument_def(
  ref,
  default_val = NA,
  required = FALSE,
  type = "character",
  nargs = 1,
  help = NA
)
}
\arguments{
\item{ref}{character:
command line option (prefix '--') or position (no prefix) argument name}

\item{default_val}{(NA | double | integer | logical | character)
(default NA)}

\item{required}{logical: 
whether the argument is required input (default FALSE)}

\item{type}{character:
Type include: character | integer | double | logical, (default character)}

\item{nargs}{character | numeric :
* integer: required number of space separated values
* "+": at least one argument
* default 1}

\item{help}{character:
documentation to be printed with --help (default NA)}
}
\value{
list
 with correct key value pairs for parsing
}
\description{
Specifies which command line options the script / program is able
     to accept and the expected properties of each option value.
}
