% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparse.R
\name{argument_def}
\alias{argument_def}
\title{Specify command line argument}
\usage{
argument_def(
  ref,
  default_val = NULL,
  required = FALSE,
  type = "character",
  nargs = NULL,
  help = NULL
)
}
\arguments{
\item{ref}{(character)
command line option (prefix '--') or position (no prefix) argument name}

\item{default_val}{(NULL | double | integer | logical | character)
(default NULL)}

\item{required}{(logical) 
(default FALSE) whether the argument is required input}

\item{type}{(character)
(default "character") can take values (character | integer | double | logical)}

\item{nargs}{(character | numeric | NULL)
(default NULL, a single option value) If:
      - integer: required number of space separated values
      - "+": at least one argument}

\item{help}{(character)
(default NULL) documentation to be printed with --help}
}
\value{
(list) with correct key value pairs for parsing
}
\description{
Specifies which command line options the script / program is able
     to accept and the expected properties of each option value.
}
